using removeprefix and rename typingdata to typinganndata
